refactor to keep import classes in a different file
